requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
moviepy>=1.0.3
asyncio>=3.4.3 